# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "BackendName": {
        "enum": [
          "pytorch",
          "mlx",
          "onnx"
        ],
        "title": "BackendName",
        "type": "string"
      },
      "BatchRequest": {
        "additionalProperties": false,
        "properties": {
          "requests": {
            "items": {
              "$ref": "#/components/schemas/InferenceRequest"
            },
            "maxItems": 256,
            "minItems": 1,
            "title": "Requests",
            "type": "array"
          }
        },
        "required": [
          "requests"
        ],
        "title": "BatchRequest",
        "type": "object"
      },
      "BatchResponse": {
        "additionalProperties": false,
        "properties": {
          "responses": {
            "items": {
              "$ref": "#/components/schemas/InferenceResponse"
            },
            "title": "Responses",
            "type": "array"
          }
        },
        "required": [
          "responses"
        ],
        "title": "BatchResponse",
        "type": "object"
      },
      "ClassificationSchema": {
        "additionalProperties": false,
        "properties": {
          "constraints": {
            "default": [],
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "title": "Constraints",
            "type": "array"
          },
          "kind": {
            "const": "classification",
            "default": "classification",
            "title": "Kind",
            "type": "string"
          },
          "tasks": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ClassificationTask"
            },
            "minProperties": 1,
            "title": "Tasks",
            "type": "object"
          }
        },
        "required": [
          "tasks"
        ],
        "title": "ClassificationSchema",
        "type": "object"
      },
      "ClassificationTask": {
        "additionalProperties": false,
        "properties": {
          "activation": {
            "default": "auto",
            "enum": [
              "auto",
              "sigmoid",
              "softmax"
            ],
            "title": "Activation",
            "type": "string"
          },
          "default": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default"
          },
          "instruction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instruction"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Labels",
            "type": "array"
          },
          "max_labels": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 1,
            "title": "Max Labels"
          },
          "min_labels": {
            "default": 1,
            "minimum": 0.0,
            "title": "Min Labels",
            "type": "integer"
          },
          "ordered": {
            "default": false,
            "title": "Ordered",
            "type": "boolean"
          },
          "temperature": {
            "default": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Temperature",
            "type": "number"
          },
          "threshold": {
            "default": 0.5,
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Threshold",
            "type": "number"
          }
        },
        "required": [
          "labels"
        ],
        "title": "ClassificationTask",
        "type": "object"
      },
      "DownloadJob": {
        "additionalProperties": false,
        "properties": {
          "destination": {
            "title": "Destination",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/DownloadState"
          }
        },
        "required": [
          "job_id",
          "model_id",
          "revision",
          "destination",
          "state"
        ],
        "title": "DownloadJob",
        "type": "object"
      },
      "DownloadRequest": {
        "additionalProperties": false,
        "properties": {
          "approved": {
            "title": "Approved",
            "type": "boolean"
          },
          "destination": {
            "minLength": 1,
            "title": "Destination",
            "type": "string"
          },
          "revision": {
            "anyOf": [
              {
                "pattern": "^[a-f0-9]{40}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          }
        },
        "required": [
          "approved",
          "destination"
        ],
        "title": "DownloadRequest",
        "type": "object"
      },
      "DownloadState": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed"
        ],
        "title": "DownloadState",
        "type": "string"
      },
      "EntitySchema": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "entities",
            "default": "entities",
            "title": "Kind",
            "type": "string"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Labels",
            "type": "array"
          }
        },
        "required": [
          "labels"
        ],
        "title": "EntitySchema",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InferenceError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "InferenceError",
        "type": "object"
      },
      "InferenceOperation": {
        "enum": [
          "classify",
          "extract_entities",
          "extract_relations",
          "extract_structured"
        ],
        "title": "InferenceOperation",
        "type": "string"
      },
      "InferenceOptions": {
        "additionalProperties": false,
        "properties": {
          "include_confidence": {
            "default": true,
            "title": "Include Confidence",
            "type": "boolean"
          },
          "output_format": {
            "$ref": "#/components/schemas/OutputFormat",
            "default": "native"
          },
          "threshold": {
            "default": 0.5,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Threshold",
            "type": "number"
          }
        },
        "title": "InferenceOptions",
        "type": "object"
      },
      "InferenceRequest": {
        "additionalProperties": false,
        "properties": {
          "backend": {
            "$ref": "#/components/schemas/BackendName"
          },
          "model": {
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "operation": {
            "$ref": "#/components/schemas/InferenceOperation"
          },
          "options": {
            "$ref": "#/components/schemas/InferenceOptions"
          },
          "precision": {
            "$ref": "#/components/schemas/Precision"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "schema": {
            "discriminator": {
              "mapping": {
                "classification": "#/components/schemas/ClassificationSchema",
                "entities": "#/components/schemas/EntitySchema",
                "relations": "#/components/schemas/RelationSchema",
                "structured": "#/components/schemas/StructuredSchema"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/ClassificationSchema"
              },
              {
                "$ref": "#/components/schemas/EntitySchema"
              },
              {
                "$ref": "#/components/schemas/RelationSchema"
              },
              {
                "$ref": "#/components/schemas/StructuredSchema"
              }
            ],
            "title": "Schema"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "model",
          "backend",
          "precision",
          "operation",
          "text",
          "schema"
        ],
        "title": "InferenceRequest",
        "type": "object"
      },
      "InferenceResponse": {
        "additionalProperties": false,
        "properties": {
          "backend": {
            "$ref": "#/components/schemas/BackendName"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InferenceError"
              },
              {
                "type": "null"
              }
            ]
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JsonValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "precision": {
            "$ref": "#/components/schemas/Precision"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "timing": {
            "$ref": "#/components/schemas/Timing"
          }
        },
        "required": [
          "request_id",
          "model",
          "backend",
          "precision",
          "timing"
        ],
        "title": "InferenceResponse",
        "type": "object"
      },
      "JsonValue": {},
...
      "OutputFormat": {
        "enum": [
          "native"
        ],
        "title": "OutputFormat",
        "type": "string"
      },
      "Precision": {
        "enum": [
          "fp32",
          "fp16",
          "bf16",
          "int8",
          "int4"
        ],
        "title": "Precision",
        "type": "string"
      },
...
      "RelationSchema": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "relations",
            "default": "relations",
            "title": "Kind",
            "type": "string"
          },
          "relation_types": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Relation Types",
            "type": "array"
          }
        },
        "required": [
          "relation_types"
        ],
        "title": "RelationSchema",
        "type": "object"
      },
      "StructuredSchema": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "structured",
            "default": "structured",
            "title": "Kind",
            "type": "string"
          },
          "schema": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "minProperties": 1,
            "title": "Schema",
            "type": "object"
          }
        },
        "required": [
          "schema"
        ],
        "title": "StructuredSchema",
        "type": "object"
      },
      "Timing": {
        "additionalProperties": false,
        "properties": {
          "inference_ms": {
            "minimum": 0.0,
            "title": "Inference Ms",
            "type": "number"
          },
          "queue_ms": {
            "minimum": 0.0,
            "title": "Queue Ms",
            "type": "number"
          }
        },
        "required": [
          "queue_ms",
          "inference_ms"
        ],
        "title": "Timing",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Local, explicitly configured GLiNER inference",
    "title": "GLiNER Runner",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/batch": {
      "post": {
        "operationId": "batch_v1_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Unsupported explicit configuration"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Queue capacity exhausted"
          }
        },
        "summary": "Batch",
        "tags": [
          "inference"
        ]
      }
    },
...
    "/v1/infer": {
      "post": {
        "operationId": "infer_v1_infer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Unsupported explicit configuration"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Queue capacity exhausted"
          }
        },
        "summary": "Infer",
        "tags": [
          "inference"
        ]
      }
    },
    "/v1/model-downloads/{job_id}": {
      "get": {
        "operationId": "download_status_v1_model_downloads__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Status",
        "tags": [
          "models"
        ]
      }
    },
...
    "/v1/models/{model_id}/downloads": {
      "post": {
        "operationId": "create_download_v1_models__model_id__downloads_post",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Download",
        "tags": [
          "models"
        ]
      }
    }
  }
}